Adding 'name' parameter to Server, which is used to find the right section of the configuration file as well as finding the pid file.
Adding 'clone' command to clone a server. Currently
uses only the naive method of backing up a server
by copying the database files.

Fixing Master role to imbue correctly event when
it cannot write the configuration file. In this
case, it is assumed that the Master was configured
correctly already.